# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ArticleStatus = (typeof ArticleStatus)[keyof typeof ArticleStatus];

export const ArticleStatus = {
  queued: "queued",
  researching: "researching",
  writing: "writing",
  humanizing: "humanizing",
  checking: "checking",
  retrying: "retrying",
  formatting: "formatting",
  completed: "completed",
  failed: "failed",
  flagged: "flagged",
} as const;

export interface Article {
  id: number;
  /** @nullable */
  title?: string | null;
  topic: string;
  primaryKeyword: string;
  /** @nullable */
  secondaryKeywords?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  /** @nullable */
  referenceInput?: string | null;
  wordCountTarget: number;
  /** @nullable */
  wordCountActual?: number | null;
  status: ArticleStatus;
  /** @nullable */
  copyleaksScore?: number | null;
  /** @nullable */
  burstinessScore?: number | null;
  /** @nullable */
  lexicalFingerprintScore?: number | null;
  /** @nullable */
  primaryKeywordDensity?: number | null;
  /** @nullable */
  secondaryKeywordDensity?: number | null;
  /** @nullable */
  emDashCount?: number | null;
  /** @nullable */
  faqCount?: number | null;
  /** @nullable */
  googleDocUrl?: string | null;
  /** @nullable */
  googleDocFileName?: string | null;
  /** @nullable */
  seoMetaDescription?: string | null;
  /** @nullable */
  seoSlug?: string | null;
  /** @nullable */
  seoTags?: string | null;
  retryCount: number;
  aiSignatureRetryCount: number;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  articleContent?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface CreateArticleBody {
  topic: string;
  primaryKeyword: string;
  secondaryKeywords?: string;
  targetAudience?: string;
  referenceInput?: string;
  wordCountTarget: number;
  createdBy?: string;
}

export interface BatchCreateBody {
  articles: CreateArticleBody[];
}

export type PipelineLogStatus =
  (typeof PipelineLogStatus)[keyof typeof PipelineLogStatus];

export const PipelineLogStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface PipelineLog {
  id: number;
  articleId: number;
  stepName: string;
  status: PipelineLogStatus;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalArticles: number;
  completedArticles: number;
  failedArticles: number;
  activeArticles: number;
  /** @nullable */
  avgCopyleaksScore?: number | null;
  /** @nullable */
  avgPrimaryDensity?: number | null;
  /** @nullable */
  successRate?: number | null;
  articlesThisWeek: number;
}

export type ListArticlesParams = {
  status?: string;
  limit?: number;
  offset?: number;
};