# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleStatus } from "./articleStatus";

export interface Article {
  id: number;
  /** @nullable */
  title?: string | null;
  topic: string;
  primaryKeyword: string;
  /** @nullable */
  secondaryKeywords?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  /** @nullable */
  referenceInput?: string | null;
  wordCountTarget: number;
  /** @nullable */
  wordCountActual?: number | null;
  status: ArticleStatus;
  /** @nullable */
  copyleaksScore?: number | null;
  /** @nullable */
  burstinessScore?: number | null;
  /** @nullable */
  lexicalFingerprintScore?: number | null;
  /** @nullable */
  primaryKeywordDensity?: number | null;
  /** @nullable */
  secondaryKeywordDensity?: number | null;
  /** @nullable */
  emDashCount?: number | null;
  /** @nullable */
  faqCount?: number | null;
  /** @nullable */
  googleDocUrl?: string | null;
  /** @nullable */
  googleDocFileName?: string | null;
  /** @nullable */
  seoMetaDescription?: string | null;
  /** @nullable */
  seoSlug?: string | null;
  /** @nullable */
  seoTags?: string | null;
  retryCount: number;
  aiSignatureRetryCount: number;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  articleContent?: string | null;
  createdAt: Date;
  /** @nullable */
  completedAt?: Date | null;
}