# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArticleBody {
  topic: string;
  primaryKeyword: string;
  secondaryKeywords?: string;
  targetAudience?: string;
  referenceInput?: string;
  wordCountTarget: number;
  createdBy?: string;
}