# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalArticles: number;
  completedArticles: number;
  failedArticles: number;
  activeArticles: number;
  /** @nullable */
  avgCopyleaksScore?: number | null;
  /** @nullable */
  avgPrimaryDensity?: number | null;
  /** @nullable */
  successRate?: number | null;
  articlesThisWeek: number;
}